[microgpt] Refactor CLI into train/infer modules
Split main.rs into three files:
- main.rs: CLI parsing and dispatch
- train.rs: unified training loop via TrainingData trait
- infer.rs: generate, chat, and info commands

Eliminates ~100 lines of duplication between the old run_train
and run_train_chat functions.
